fix: pnl calculations

+++ src/common/transfers.ts
@@ -87,10 +87,11 @@ export function convertValueToUnderlying(
         if (activeMarkets[i].maturity == (token.maturity as BigInt)) {
           let lastImpliedRate = activeMarkets[i].lastImpliedRate;
           let timeToMaturity = (token.maturity as BigInt).minus(blockTime);
-          let x: f64 = (lastImpliedRate
-            .times(timeToMaturity)
-            .div(RATE_PRECISION)
-            .toI64() / SECONDS_IN_YEAR.toI64()) as f64;
+          let x: f64 =
+            (lastImpliedRate
+              .times(timeToMaturity)
+              .div(RATE_PRECISION)
+              .toI64() as f64) / (SECONDS_IN_YEAR.toI64() as f64);
 
           let discountFactor = BigInt.fromI64(
             Math.floor(Math.exp(-x) * (RATE_PRECISION.toI64() as f64)) as i64
